pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voxpop"
description = ''
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = []
authors = [
  { name = "Semaphor", email = "info@semaphor.dk" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "Django>=4.2",
  "django-ninja==0.21.0",
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/semaphor-dk/voxpop#readme"
Issues = "https://github.com/semaphor-dk/voxpop/issues"
Source = "https://github.com/semaphor-dk/voxpop/"

[tool.hatch.version]
path = "voxpop/__about__.py"

[tool.hatch.build.targets.sdist]
include = [
    "LICENSE.txt",
    "README.txt",
    "/voxpop",
]

[tool.hatch.envs.default]
dependencies = [
    "coverage[toml]==7.2.1",
    "pytest==7.2.2",
    "pytest-cov",
    "pytest-django==4.5.2",
    "mypy==1.1.1",
    "django-stubs==1.16.0",
]

[tool.hatch.envs.default.scripts]
cov = "pytest -rsx --cov-report=term-missing --cov-config=pyproject.toml --cov=voxpop --cov=tests --cov=append {args}"
no-cov = "cov --no-cov {args}"
typecheck = "mypy --config-file=pyproject.toml ."

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE="tests.settings"
addopts = "--reuse-db"
norecursedirs = "build dist docs .eggs/* *.egg-info htmlcov voxpop .git"
python_files = "test*.py"
testpaths = "tests"
pythonpath = ". tests"

[tool.coverage.run]
branch = true
parallel = true
omit = [
    "voxpop/__about__.py",
    "tests/manage.py",
]

[tool.coverage.report]
exclude_lines = [
    "no cov",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.mypy]
mypy_path = "voxpop/"
exclude = [
    "venv/",
    "dist/",
    "docs/",
]
namespace_packages = false
show_error_codes = true
strict = true
warn_unreachable = true
follow_imports = "normal"

[tool.django-stubs]
#django_settings_module = "tests.settings"

[[tool.mypy.overrides]]
module = "tests.*"
allow_untyped_defs = true
